[all (rust)] Pull out structures for intermediate representation
Also fix the data type for thin timestamp
