+++ src/cache/mod.rs
@@ -1,6 +1,7 @@
 pub mod check;
 pub mod dump;
 pub mod hint;
+pub mod ir;
 pub mod mapping;
 pub mod restore;
 pub mod superblock;
