+++ src/thin/mod.rs
@@ -2,6 +2,7 @@ pub mod block_time;
 pub mod check;
 pub mod device_detail;
 pub mod dump;
+pub mod ir;
 pub mod restore;
 pub mod runs;
 pub mod superblock;
